requirements.txt
asn1crypto==0.22.0
chardet==3.0.4
colorama==0.3.7
cryptography==1.9
decorator==4.1.1
deepin-menu==0.1
Django==1.11.13
djangorestframework==3.8.2
dnspython==1.15.0
enum34==1.1.6
idna==2.5
ipaddress==1.0.17
keyring==10.4.0
keyrings.alt==2.2
mysql-connector-python==2.1.6
mysql-utilities==1.6.4
numpy==1.12.1
paramiko==2.0.0
pathlib2==2.3.0
pbr==4.0.2
pexpect==4.2.1
prettytable==0.7.2
psutil==5.0.1
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pyodbc==3.0.10
pyOpenSSL==16.2.0
pysmbc==1.0.15.6
pysqlite==2.7.0
pytz==2018.4
pyxdg==0.25
requests==2.12.4
scandir==1.5
SecretStorage==2.3.1
six==1.10.0
stevedore==1.28.0
urllib3==1.19.1
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
